reports

+++ app/Console/Commands/UploadAssetReport.php
@@ -3,6 +3,7 @@
 namespace App\Console\Commands;
 
 use App\Jobs\HousingPage;
+use App\Models\ParsingConstant;
 use App\Models\AssetReport;
 use App\Models\AssetReportValidation;
 use App\Service\HousingUploader;
@@ -37,7 +38,22 @@ class UploadAssetReport extends Command
      */
     public function handle()
     {
-        AssetReport::query()->truncate();;
+        $assetConstant = ParsingConstant::firstOrCreate(
+            [
+                'type' => ParsingConstant::ASSET_TYPE,
+            ],
+            [
+                'is_need_parsing' => false,
+            ]
+        );
+
+        if (!$assetConstant->is_need_parsing) {
+            return;
+        }
+        $assetConstant->is_need_parsing = false;
+        $assetConstant->save();
+
+        AssetReport::query()->truncate();
         AssetReportValidation::query()->truncate();
         $pages = $this->simpro->getRequestPage('get', self::SITES_URL.'?Customers.ID=11514&columns=ID,CustomFields');
         foreach ($pages as $page) {
@@ -131,7 +147,7 @@ class UploadAssetReport extends Command
             $data['job_due_date'] = $job->DueDate ?? null;
 
             $jobId = $job->ID ?? 0;
-                $job = $this->simpro->getRequest('get', '/api/v1.0/companies/0/jobs/'.$jobId.'?columns=Stage,Tags,DueDate');
+            $job = $this->simpro->getRequest('get', '/api/v1.0/companies/0/jobs/'.$jobId.'?columns=Stage,Tags,DueDate');
             if ($job) {
                 $data['job_stage'] = $job->Stage ?? null;
                 $data['service_due'] = $job->DueDate;
@@ -147,7 +163,7 @@ class UploadAssetReport extends Command
                 }
             }
 
-                $schedules = $this->simpro->getRequest('get', '/api/v1.0/companies/0/schedules/?Reference='.$jobId.'-%');
+            $schedules = $this->simpro->getRequest('get', '/api/v1.0/companies/0/schedules/?Reference='.$jobId.'-%');
             if ($schedules) {
                 $schedule = $schedules[0] ?? new \stdClass();
                 $scheduleDate = $schedule->Date;
